feat/update_condition+price
